lastFiles: ENH: Allow passing FILESPECs (e.g. through glob) as well, and consider the globally last filename when multiple DIRSPECs are passed
So far, the behavior was inconsistent with the sister newestFiles command; the price for a simple implementation that did not rely on any other external commands, and relied on the built-in ordering of the glob result.
But as the ability to pass globs is useful (e.g. to determine the last PDF file by passing path/to/*.pdf), let's bite the bullet and do a complete implementation that first collects all filespecs with prepended filenames, then sorts that list, strips off the prepending filenames, and then indexes into that list.