touch-p: Also mention created filespecs when -v|--verbose is given
Just because touch doesn't support this natively like mkdir does not mean that this information isn't as important; especially when both directories and the file are missing, it's important to have the complete picture.
Restructure the command-line argument parsing to avoid duplicating the isVerbose flag setting, and this also gets rid of the extglob setting.