writeorig: Make --keep change the exit status to 0 if the backup already exists
The option so far was more or less useless, as keeping the FILE.orig is the default; however, the failure to write a backup was still indicated by exiting with 1. My pipethrough[1] command can now take a backup command, and will abort if the file cannot be backed up. To make that place nicely with writeorig, we can now use the --keep argument which turns the exit status into unconditional success (well, except for truly exceptional conditions), so we don't have to resort to tricks like "acceptStatus 0 1 writeorig" (which would also suppress those exceptional conditions).

Note: writeOrigOrBackup needs to drop the (formerly superfluous) --keep argument, because it relies on writeorig's failure to then switch to writebackup.