Documentation: Highlight that these are about FILEGLOBs
There are subtle differences in the glob that gets evaluated by a switch statement (can wrap in @(...) even if there's no globbing character, works across / boundary) and a real file glob. For the latter, we really need to use eval, and yet special shell characters (that had been properly escaped) are fine.