expandglob: BUG: Whitespace in FILE-GLOB prevents match
Whitespace within a FILE[-GLOB] argument should belong to the file and must not need escaping. However, the eval used in the globbing leaves whitespace unsecured, so each word is parsed as a separate glob, which is wrong.
As we still want the glob characters to be active, we cannot %q-quote, but instead just manually escape spaces, and through IFS=' ' make that the only word separator (so we don't have to deal with newlines and tabs).