ENH: Offer recursive deletion of empty parent directories via -p|--parents
rmdir has this as well; however it only deletes those path components that are actually passed. The value-add of rmemptydir is that it resolves each passed DIR to its absolute dirspec, so any empty parent directories are removed.
This fits well into the original use case of deleting empty child directories; now it's multi-directional.