ENH: Influence the directory handling via -d|--include-directories|-r|--recursive
By default, directories are considered as themselves; this is inconsistent with newestFile, which considers the whole directory tree (as it's find-based, but there this can be stopped via -maxdepth 1).
Let's instead ignore directories by default, and offer the previous behavior under -d|--include-directories. An added complication here is that DIRSPEC can both mean a starting dir (and the CWD default internally is handled like this, too), or one directory in a list of passed files. For the latter, we want filtering from the list, but for the former we want to look inside that directory. To be able to differentiate between the two, add the rule that the former only applies when a single DIRSPEC is passed.
Then there's the -r|--recursive case that makes it behave like newestFile.